refactor: Merge outgoing request usecases with protocol call usecases
