+++ sync/sdk/src/main/kotlin/com/walletconnect/sync/engine/use_case/calls/DeleteStoreValueUseCase.kt
@@ -1,18 +1,35 @@
 package com.walletconnect.sync.engine.use_case.calls
 
 import com.walletconnect.android.internal.common.model.AccountId
+import com.walletconnect.android.internal.common.model.IrnParams
+import com.walletconnect.android.internal.common.model.Tags
+import com.walletconnect.android.internal.common.model.type.JsonRpcInteractorInterface
 import com.walletconnect.android.internal.common.scope
+import com.walletconnect.android.internal.utils.MONTH_IN_SECONDS
+import com.walletconnect.foundation.common.model.Topic
+import com.walletconnect.foundation.common.model.Ttl
 import com.walletconnect.sync.common.exception.validateAccountId
+import com.walletconnect.sync.common.json_rpc.SyncParams
+import com.walletconnect.sync.common.json_rpc.SyncRpc
 import com.walletconnect.sync.common.model.Store
-import com.walletconnect.sync.engine.use_case.requests.outgoing.SendDeleteRequestUseCase
 import com.walletconnect.sync.storage.StoresStorageRepository
+import com.walletconnect.util.generateId
 import kotlinx.coroutines.launch
 import kotlinx.coroutines.supervisorScope
 
-internal class DeleteStoreValueUseCase(private val storesRepository: StoresStorageRepository, private val sendDeleteRequestUseCase: SendDeleteRequestUseCase) : DeleteUseCaseInterface {
+internal class DeleteStoreValueUseCase(private val storesRepository: StoresStorageRepository, private val jsonRpcInteractor: JsonRpcInteractorInterface) : DeleteUseCaseInterface {
 
     // https://github.com/WalletConnect/WalletConnectKotlinV2/issues/800 -> update params to have StoreKey and StoreValue
     override fun delete(accountId: AccountId, store: Store, key: String, onSuccess: (Boolean) -> Unit, onFailure: (Throwable) -> Unit) {
+        suspend fun publishDeleteRequest() {
+            val storeTopic = runCatching { storesRepository.getStoreTopic(accountId, store) }.fold(onSuccess = { Topic(it) }, onFailure = { error -> return onFailure(error) })
+            val deleteParams = SyncParams.DeleteParams(key)
+            val payload = SyncRpc.SyncDelete(id = generateId(), params = deleteParams)
+            val irnParams = IrnParams(Tags.SYNC_DELETE, Ttl(MONTH_IN_SECONDS))
+
+            jsonRpcInteractor.publishJsonRpcRequest(storeTopic, irnParams, payload, onSuccess = { onSuccess(true) }, onFailure = onFailure)
+        }
+
         scope.launch {
             supervisorScope {
                 validateAccountId(accountId) { error -> return@supervisorScope onFailure(error) }
@@ -22,7 +39,7 @@ internal class DeleteStoreValueUseCase(private val storesRepository: StoresStora
 
                 // Return true in onSuccess when the value was deleted
                 runCatching { storesRepository.deleteStoreValue(accountId, store, key) }.fold(
-                    onSuccess = { sendDeleteRequestUseCase(key, accountId, store, onSuccess = { onSuccess(true) }, onError = onFailure) },
+                    onSuccess = { publishDeleteRequest() },
                     onFailure = { error -> onFailure(error) }
                 )
             }
