+++ sync/sdk/src/main/kotlin/com/walletconnect/sync/engine/use_case/requests/OnSetRequestUseCase.kt
@@ -1,4 +1,4 @@
-package com.walletconnect.sync.engine.use_case.requests.incoming
+package com.walletconnect.sync.engine.use_case.requests
 
 import com.walletconnect.android.internal.common.model.WCRequest
 import com.walletconnect.sync.common.json_rpc.SyncParams
